feat: add reward classifiers for alignment and color, including full training pipelines and dataset management
- Updated .gitignore to exclude data directories
- Enhanced README.md with detailed usage instructions for reward classifiers
- Introduced the reward_models package with dedicated classifiers for alignment and color tasks
- Implemented custom datasets and end-to-end training loops for both classifiers
- Added HSV label generation utilities and supporting scripts for color classification
- Created shared utility modules for unified dataset handling and classifier training
- Integrated necessary imports and configurations for ResNet-based classifier backbones
