+++ rl_garden/models/reward/__init__.py
@@ -1,27 +1,25 @@
-"""Classifier-specific helpers for reward models."""
+"""Reward classifier datasets, networks, and training helpers."""
 
-from rl_garden.reward_models.classifiers.alignment import (
+from rl_garden.models.reward.classifiers import (
     AlignmentClassifierConfig,
     AlignmentClassifierTrainer,
     AlignmentRewardDataset,
-    CombinedAlignmentDataset,
-)
-from rl_garden.reward_models.classifiers.color import (
     ColorClassifierConfig,
     ColorClassifierTrainer,
     ColorRewardDataset,
+    CombinedAlignmentDataset,
     CombinedColorRewardDataset,
+    ResNetBinaryClassifier,
 )
-from rl_garden.reward_models.classifiers.shared.model import ResNetBinaryClassifier
 
 __all__ = [
     "AlignmentClassifierConfig",
     "AlignmentClassifierTrainer",
     "AlignmentRewardDataset",
-    "CombinedAlignmentDataset",
     "ColorClassifierConfig",
     "ColorClassifierTrainer",
     "ColorRewardDataset",
+    "CombinedAlignmentDataset",
     "CombinedColorRewardDataset",
     "ResNetBinaryClassifier",
 ]
