+++ rl_garden/models/reward/classifiers/color/dataset.py
@@ -9,8 +9,8 @@ import numpy as np
 import torch
 from torch.utils.data import ConcatDataset, Dataset
 
-from rl_garden.reward_models.classifiers.shared.io import decode_jpeg_crop
-from rl_garden.reward_models.classifiers.shared.transforms import build_transforms, empty_image
+from rl_garden.models.reward.classifiers.shared.io import decode_jpeg_crop
+from rl_garden.models.reward.classifiers.shared.transforms import build_transforms, empty_image
 
 
 class ColorRewardDataset(Dataset):
