+++ rl_garden/models/reward/classifiers/shared/model.py
